binded numeric insert workaround (no implementation in rust-postgres)
The "postgres" library requires binded variables to be of a rust type
that has their "ToSql" trait implemented. there exist no viable types
that implement this trait for the Postgres numeric type. "decimal"
package provides an implementation, but this package's type only goes
up to 96 bits size values.

We should implement ToSql for for example crate::num's BigInt, or for
crate::big_decimal's BigDecimal. But until we've done that, this commit
enables inserting numerics through a bit of a hack in the insert
statement.